made array of objects in main.cpp
